pyproject.toml
[project]
name = "lightweight-etl-pipeline-to-gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.4.0",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-storage>=3.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "setuptools>=80.9.0",
]